+++ beatmap/v3/lightTranslationEventBox.ts
@@ -44,18 +44,10 @@ export class LightTranslationEventBox extends WrapLightTranslationEventBox<
 
     private _f: IndexFilter;
     private _l: LightTranslationBase[];
-    protected constructor(
-        lightTranslationEventBox: Required<ILightTranslationEventBox>,
-    ) {
+    protected constructor(lightTranslationEventBox: Required<ILightTranslationEventBox>) {
         super(lightTranslationEventBox);
         this._f = IndexFilter.create(lightTranslationEventBox.f);
         this._l = lightTranslationEventBox.l.map((l) => LightTranslationBase.create(l)[0]);
-        const lastTime = Math.max(...this._l.map((l) => l.time));
-        if (this.beatDistributionType === 2) {
-            this.beatDistribution = this.beatDistribution < lastTime
-                ? lastTime
-                : this.beatDistribution;
-        }
     }
 
     static create(): LightTranslationEventBox[];
@@ -102,24 +94,18 @@ export class LightTranslationEventBox extends WrapLightTranslationEventBox<
                     f: (eb.filter as IIndexFilter) ??
                         (eb as Required<ILightTranslationEventBox>).f ??
                         LightTranslationEventBox.default.f(),
-                    w: eb.beatDistribution ?? eb.w ??
-                        LightTranslationEventBox.default.w,
-                    d: eb.beatDistributionType ?? eb.d ??
-                        LightTranslationEventBox.default.d,
-                    s: eb.translationDistribution ?? eb.s ??
-                        LightTranslationEventBox.default.s,
-                    t: eb.translationDistributionType ?? eb.t ??
-                        LightTranslationEventBox.default.t,
+                    w: eb.beatDistribution ?? eb.w ?? LightTranslationEventBox.default.w,
+                    d: eb.beatDistributionType ?? eb.d ?? LightTranslationEventBox.default.d,
+                    s: eb.translationDistribution ?? eb.s ?? LightTranslationEventBox.default.s,
+                    t: eb.translationDistributionType ?? eb.t ?? LightTranslationEventBox.default.t,
                     a: eb.axis ?? eb.a ?? LightTranslationEventBox.default.a,
                     r: eb.flip ?? eb.r ?? LightTranslationEventBox.default.r,
-                    b: eb.affectFirst ?? eb.b ??
-                        LightTranslationEventBox.default.b,
+                    b: eb.affectFirst ?? eb.b ?? LightTranslationEventBox.default.b,
                     i: eb.easing ?? eb.i ?? LightTranslationEventBox.default.i,
                     l: (eb.events as ILightTranslationBase[]) ??
                         (eb as Required<ILightTranslationEventBox>).l ??
                         LightTranslationEventBox.default.l(),
-                    customData: eb.customData ??
-                        LightTranslationEventBox.default.customData(),
+                    customData: eb.customData ?? LightTranslationEventBox.default.customData(),
                 }),
             )
         );
@@ -232,9 +218,7 @@ export class LightTranslationEventBox extends WrapLightTranslationEventBox<
     get customData(): NonNullable<ILightTranslationEventBox['customData']> {
         return this.data.customData;
     }
-    set customData(
-        value: NonNullable<ILightTranslationEventBox['customData']>,
-    ) {
+    set customData(value: NonNullable<ILightTranslationEventBox['customData']>) {
         this.data.customData = value;
     }